Added documentation for all important functions
