+++ .gitignore
@@ -116,4 +116,3 @@ Temporary Items
 .vscode
 .venv
 output/
-shell.nix
