+++ Makefile
@@ -6,7 +6,7 @@ OUT_DIR = output
 TOK_DIR = src/parsing/tokenisation
 AST_DIR = src/parsing/ast
 MINI_AST_SRC = $(AST_DIR)/parse.c
-MINI_TOKENISATION_SRC = $(TOK_DIR)/tokens.c $(TOK_DIR)/single_token_utils.c $(TOK_DIR)/token_list_utils.c $(TOK_DIR)/tokens_handlers.c
+MINI_TOKENISATION_SRC = $(TOK_DIR)/tokens.c $(TOK_DIR)/single_token_utils.c $(TOK_DIR)/token_list_utils.c $(TOK_DIR)/tokens_handlers.c $(TOK_DIR)/tokens_handler_01.c 
 MINI_SRC = $(MINI_TOKENISATION_SRC) $(MINI_AST_SRC)
 MAIN_SRC = src/main.c $(MINI_SRC)
 MAIN_OBJ = $(MAIN_SRC:.c=.o)
